refactor: 肥大コンポーネント4件を責務ごとに分割（Phase 2）
MyReservations と同じパターン（派生ロジックはフック・大きなJSXはサブコンポーネント）で分割:
- BookingWizard: use-availability-groups / use-booking-cart / use-booking-submit(+test) / StepIndicator
- CalendarBoard: use-calendar-view-state / use-calendar-derived(+test) / CalendarHeader / MonthView / GanttView
- settings/page: use-profile-actions / ProfileSection / settings-primitives
- UserMenu: FeedbackDialog / ManagerKeyDialog
振る舞いは不変（tsc・build・test 435件すべて green）。
