+++ app/(protected)/ems/equipment-list/_components/BookingWizard.tsx
@@ -2,7 +2,6 @@
 
 import { useMemo, useState } from "react";
 import { useRouter } from "next/navigation";
-import { toast } from "sonner";
 import { cn } from "@/lib/utils";
 
 import { useCachedEndpoint } from "@/hooks/use-cached-endpoint";
@@ -10,26 +9,24 @@ import { useMonthNav } from "@/hooks/use-month-nav";
 import { useEquipments } from "@/app/(protected)/ems/hooks/use-equipments";
 import { useCategories } from "@/app/(protected)/ems/equipment-list/hooks/use-categories";
 import { useReserves } from "@/app/(protected)/ems/equipment-list/hooks/use-reserves";
-import { useCreateReservations } from "@/app/(protected)/ems/equipment-list/hooks/use-create-reservations";
+import { useAvailabilityGroups } from "@/app/(protected)/ems/equipment-list/hooks/use-availability-groups";
+import { useBookingCart } from "@/app/(protected)/ems/equipment-list/hooks/use-booking-cart";
+import { useBookingSubmit } from "@/app/(protected)/ems/equipment-list/hooks/use-booking-submit";
 
 import {
   buildMonthMatrix,
-  toJstDayIndex,
   todayJstDayIndex,
   formatRange,
   dayIndexToDateString,
 } from "@/lib/calendar/date-grid";
-import { categoryColor, categoryIconPath } from "@/lib/category-colors";
-import { flattenNewlines } from "@/lib/text";
 import type { DayRange } from "@/components/calendar/RangeMiniCalendar";
-import type { Equipment } from "@/types/domain";
 import { Skeleton } from "@/components/ui/skeleton";
 
 import { PeriodPanel } from "./PeriodPanel";
 import { EquipmentPickPanel } from "./EquipmentPickPanel";
 import { ConfirmPanel } from "./ConfirmPanel";
 import { DoneScreen } from "./DoneScreen";
-import type { CartItem, PickGroup } from "./types";
+import { StepIndicator, StepDot } from "./StepIndicator";
 
 interface UserLite {
   id: string;
@@ -47,7 +44,6 @@ export function BookingWizard() {
     hasLoaded: reservesLoaded,
     refetch: refetchReserves,
   } = useReserves();
-  const { isSubmitting, createReservations } = useCreateReservations();
 
   // 部員名の解決用。キャッシュはカレンダー画面の /api/users と共有される
   const { data: users } = useCachedEndpoint<UserLite>("/api/users");
@@ -57,7 +53,6 @@ export function BookingWizard() {
   const matrix = useMemo(() => buildMonthMatrix(viewYear, viewMonth0), [viewYear, viewMonth0]);
 
   const [range, setRange] = useState<DayRange>({ startIdx: null, endIdx: null });
-  const [cart, setCart] = useState<number[]>([]);
   const [step, setStep] = useState<1 | 2 | 3>(1); // モバイルのステップ
   const [done, setDone] = useState(false);
   // 機材が増えても目当てを探せるよう、名前検索と「空きのみ」絞り込みを持つ
@@ -70,143 +65,36 @@ export function BookingWizard() {
   const rangeText = rangeOk ? formatRange(range.startIdx!, range.endIdx!) : "";
   const days = rangeOk ? range.endIdx! - range.startIdx! + 1 : 0;
 
-  const userName = (id: string) => users.find((u) => u.id === id)?.name ?? "他の人";
   const equipmentName = (id: number) => equipments.find((e) => e.id === id)?.name ?? `#${id}`;
 
-  // カテゴリ順にグループ化し、期間中の空き状況を付与する
-  const groups = useMemo<PickGroup[]>(() => {
-    if (!rangeOk) return [];
+  const { cart, setCart, toggle, pruneForRange } = useBookingCart({ reserves, equipmentName });
 
-    const toItem = (e: Equipment) => {
-      // 競合は全件拾って開始日順に出す。先頭1件だけだと「表示された予約を避けて
-      // 期間を選び直したのにまだ弾かれる」という手戻りが起きる（DB挿入順は日付順ですらない）。
-      const conflicts = reserves
-        .filter(
-          (r) =>
-            r.list_id === e.id &&
-            range.startIdx! <= toJstDayIndex(r.end) &&
-            range.endIdx! >= toJstDayIndex(r.start)
-        )
-        .sort((a, b) => toJstDayIndex(a.start) - toJstDayIndex(b.start));
-      const free = conflicts.length === 0;
-      const first = conflicts[0];
-      return {
-        id: e.id,
-        name: e.name,
-        detail: flattenNewlines(e.detail ?? ""),
-        image: e.image ?? "",
-        free,
-        sub: free
-          ? "この期間は空いています"
-          : `${formatRange(toJstDayIndex(first.start), toJstDayIndex(first.end))} ${userName(first.user_id)}が予約${conflicts.length > 1 ? ` ほか${conflicts.length - 1}件` : ""}`,
-        selected: cart.includes(e.id),
-      };
-    };
+  const { groups, visibleGroups, cartItems } = useAvailabilityGroups({
+    equipments,
+    categories,
+    catLoading,
+    reserves,
+    users,
+    range,
+    rangeOk,
+    cart,
+    query,
+    freeOnly,
+  });
 
-    const grouped = categories.map((cat) => ({
-      catId: String(cat.id),
-      catName: cat.name,
-      color: categoryColor(cat.color),
-      iconPath: categoryIconPath(cat.name),
-      items: equipments.filter((e) => String(e.tag_id) === String(cat.id)).map(toItem),
-    }));
-
-    // カテゴリ未設定・カテゴリ削除後の機材も「未分類」として末尾に出す。
-    // 以前はどのグループにも入らず、部員からは機材が消えて予約不可能になっていた。
-    // 読み込み中は全機材が未分類に見えてしまうため isLoading で判定する
-    // （categories.length だと最後の1カテゴリを削除した後に未分類が出なくなる）。
-    if (!catLoading) {
-      const knownIds = new Set(categories.map((c) => String(c.id)));
-      const uncategorized = equipments.filter(
-        (e) => e.tag_id == null || !knownIds.has(String(e.tag_id))
-      );
-      if (uncategorized.length > 0) {
-        grouped.push({
-          catId: "uncategorized",
-          catName: "未分類",
-          color: categoryColor(null),
-          iconPath: categoryIconPath(null),
-          items: uncategorized.map(toItem),
-        });
-      }
-    }
-
-    return grouped.filter((g) => g.items.length > 0);
-  }, [categories, catLoading, equipments, reserves, rangeOk, range, cart, users]);
-
-  // 検索・空きのみ絞り込み（絞り込みで空になったカテゴリは出さない）
-  const visibleGroups = useMemo<PickGroup[]>(() => {
-    const q = query.trim().toLowerCase();
-    if (!q && !freeOnly) return groups;
-    return groups
-      .map((g) => ({
-        ...g,
-        items: g.items.filter(
-          (it) => (!freeOnly || it.free) && (!q || it.name.toLowerCase().includes(q))
-        ),
-      }))
-      .filter((g) => g.items.length > 0);
-  }, [groups, query, freeOnly]);
-
-  const cartItems = useMemo<CartItem[]>(
-    () =>
-      cart.map((id) => {
-        const e = equipments.find((q) => q.id === id);
-        const cat = categories.find((c) => String(c.id) === String(e?.tag_id));
-        const color = categoryColor(cat?.color);
-        return {
-          id,
-          name: e?.name ?? "",
-          image: e?.image ?? "",
-          color,
-          iconPath: categoryIconPath(cat?.name),
-        };
-      }),
-    [cart, equipments, categories]
-  );
-
-  const toggle = (id: number) =>
-    setCart((prev) => (prev.includes(id) ? prev.filter((x) => x !== id) : [...prev, id]));
+  const { isSubmitting, handleSubmit } = useBookingSubmit({
+    cart,
+    startStr,
+    endStr,
+    equipmentName,
+    refetchReserves,
+    setDone,
+    setCart,
+  });
 
   const resetRangeSelection = () => {
     setCart([]);
     setStep(1);
-  };
-  const handleSubmit = async () => {
-    if (cart.length === 0) return;
-    const res = await createReservations(cart, startStr, endStr);
-    if (res.ok) {
-      // 完了画面は空き状況を使わないので refetch を待たずに先へ進める
-      // （待つ間に isSubmitting が解けて確定ボタンが再活性化し、二度押しできる窓があった）。
-      setDone(true);
-      void refetchReserves();
-      return;
-    }
-    await refetchReserves();
-    if (res.createdCount > 0) {
-      // 部分成功: 予約できた機材はカートから外し、できなかった機材を機材名で明示する。
-      // 全体を失敗のように伝えると、成功分まで期間を変えて予約し直す二重予約につながる。
-      // トーストは1本に統合する（2本重ねるとモバイルでは背面の成功分が読めない）。
-      setCart((prev) => prev.filter((id) => !res.createdIds.includes(id)));
-      if (res.conflictIds.length > 0) {
-        toast.error(
-          `${res.conflictIds.map(equipmentName).join("、")} は期間の重なる予約があり予約できませんでした。それ以外の${res.createdCount}件は予約済みです（マイ予約で確認できます）。期間を変えて再度お試しください。`,
-          { duration: 10000 }
-        );
-      } else {
-        toast.error(
-          `一部の機材が予約できませんでした（${res.createdCount}件は予約済み。マイ予約で確認できます）。${res.errorMessage ?? ""}`,
-          { duration: 10000 }
-        );
-      }
-    } else if (res.conflict) {
-      toast.error(
-        `${res.conflictIds.map(equipmentName).join("、")} は選択した期間にすでに予約が入っています。期間を変更してください。`
-      );
-    } else {
-      // API が返す具体的な理由（「予約開始日は今日以降にしてください。」等）をそのまま見せる
-      toast.error(res.errorMessage ?? "予約の作成中にエラーが発生しました。");
-    }
   };
 
   const restart = () => {
@@ -271,24 +159,7 @@ export function BookingWizard() {
       range={range}
       onRangeChange={(r) => {
         setRange(r);
-        // 期間変更でカートを全消去すると「期間を変更してください」の案内に従った
-        // ユーザーが機材を全部選び直すことになるため、新しい期間で空きがない機材
-        // だけを外す（期間選択の途中 = endIdx 未確定の間は何もしない）。
-        if (r.startIdx == null || r.endIdx == null) return;
-        const dropped = cart.filter((id) =>
-          reserves.some(
-            (rv) =>
-              rv.list_id === id &&
-              r.startIdx! <= toJstDayIndex(rv.end) &&
-              r.endIdx! >= toJstDayIndex(rv.start)
-          )
-        );
-        if (dropped.length > 0) {
-          setCart((prev) => prev.filter((id) => !dropped.includes(id)));
-          toast(
-            `${dropped.map(equipmentName).join("、")} は新しい期間では空きがないため選択から外しました。`
-          );
-        }
+        pruneForRange(r);
       }}
       onPrevMonth={goPrevMonth}
       onNextMonth={goNextMonth}
@@ -435,63 +306,5 @@ export function BookingWizard() {
         )}
       </div>
     </>
-  );
-}
-
-function StepDot({ n, active }: { n: number; active?: boolean }) {
-  return (
-    <span
-      className={cn(
-        "flex h-[22px] w-[22px] items-center justify-center rounded-full text-[11.5px] font-bold",
-        active ? "bg-brand text-white" : "bg-white text-ink-faint"
-      )}
-      style={!active ? { border: "2px solid #D0D5DD" } : undefined}
-    >
-      {n}
-    </span>
-  );
-}
-
-function StepIndicator({ step }: { step: 1 | 2 | 3 }) {
-  const steps = [
-    { n: 1, label: "期間" },
-    { n: 2, label: "機材" },
-    { n: 3, label: "確認" },
-  ];
-  return (
-    <div className="mb-4 flex items-center px-2">
-      {steps.map((st, i) => {
-        const active = step === st.n;
-        const done = step > st.n;
-        return (
-          <div key={st.n} className="flex items-center" style={{ flex: i < 2 ? 1 : "none" }}>
-            <div className="flex flex-none flex-col items-center gap-1">
-              <span
-                className="flex h-[30px] w-[30px] items-center justify-center rounded-full text-[13px] font-bold transition-colors"
-                style={{
-                  background: active || done ? "#2E90FA" : "#fff",
-                  border: `2px solid ${active || done ? "#2E90FA" : "#D0D5DD"}`,
-                  color: active || done ? "#fff" : "#98A2B3",
-                }}
-              >
-                {done ? "✓" : st.n}
-              </span>
-              <span
-                className="text-[10px] font-bold"
-                style={{ color: active ? "#101828" : "#98A2B3" }}
-              >
-                {st.label}
-              </span>
-            </div>
-            {i < 2 && (
-              <div
-                className="mx-1.5 mb-4 h-0.5 flex-1"
-                style={{ background: step > st.n ? "#2E90FA" : "#E4E7EC" }}
-              />
-            )}
-          </div>
-        );
-      })}
-    </div>
   );
 }
